Day 15 Part One: Oxygen System: Intcode-driven repair droid in an unknown maze
After running around the maze randomly for ten million steps, the robot found all the walls and floors and the oxygen system. I started to research a more optimized algorithm (Dijkstra's or the simplified uniform cost search) but decided to try a random walk first, and that completed faster than implementing a more complex algorithm.